requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
annotated-types==0.7.0
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.1.4
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pypng==0.20220715.0
python-dotenv==1.2.1
qrcode==7.4.2
rsa==4.9.1
SQLAlchemy==2.0.45
Telethon==1.42.0
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
